add and delete

+++ backend/src/api/entries/index.js
@@ -42,10 +42,10 @@ module.exports = {
     }
   },
   getEntries: async (req, res) => {
-    // if (!req.user) {
-    //   res.sendStatus(401);
-    //   return;
-    // }
+    if (!req.user) {
+      res.sendStatus(401);
+      return;
+    }
 
     const data = await controller.getEntries(req.db, req.query.active);
     if (data.length === 0) {
@@ -68,10 +68,10 @@ module.exports = {
     }
   },
   deleteEntrie: async (req, res) => {
-    //if (!req.user) {
-    //res.sendStatus(401);
-    //return;
-    //}
+    if (!req.user) {
+      res.sendStatus(401);
+      return;
+    }
     try {
       await getStatus(req.db, req.params.id);
       if (await controller.deleteEntrie(req.db, req.params.id)) {
